requirements.txt
biopython>=1.83
pandas>=2.0
numpy>=1.24
seqconverter
figtree